+++ lib/kompost/kompo/postgres/instance.ex
@@ -8,11 +8,13 @@ defmodule Kompost.Kompo.Postgres.Instance do
   alias Kompost.Kompo.Postgres.ConnectionRegistry
   alias Kompost.Kompo.Postgres.ConnectionSupervisor
 
+  alias Kompost.Tools.NamespaceAccess
+
   @typedoc """
   Defines the ID for an instance. The ID used as key when registering the
   Postgrex process at the `Kompost.Kompo.Postgres.ConnectionRegistry`
   """
-  @type id :: {namespace :: binary(), name :: binary()}
+  @type id :: {namespace :: binary() | :cluster, name :: binary()}
 
   @doc """
   Checks in the `Kompost.Kompo.Postgres.ConnectionRegistry` for an existing
@@ -21,20 +23,32 @@ defmodule Kompost.Kompo.Postgres.Instance do
   the `conn_args`. The process is then registered at the
   `Kompost.Kompo.Postgres.ConnectionRegistry`.
   """
-  @spec connect(id(), conn_args :: Keyword.t()) ::
+  @spec connect(
+          id(),
+          conn_args :: Keyword.t(),
+          allowed_namespaces :: NamespaceAccess.allowed_namespaces()
+        ) ::
           {:ok, Postgrex.conn()}
           | {:error, Postgrex.Error.t() | Exception.t()}
           | DynamicSupervisor.on_start_child()
-  def connect(id, conn_args) do
+  def connect(id, conn_args, allowed_namespaces) do
     with {:lookup, []} <- {:lookup, lookup(id)},
          {:ok, _} <-
            conn_args
            |> Postgrex.Utils.default_opts()
            |> Postgrex.Protocol.connect() do
-      args = Keyword.put(conn_args, :name, {:via, Registry, {ConnectionRegistry, id, conn_args}})
+      args =
+        Keyword.put(
+          conn_args,
+          :name,
+          {:via, Registry,
+           {ConnectionRegistry, id,
+            Keyword.put(conn_args, :allowed_namespaces, allowed_namespaces)}}
+        )
+
       DynamicSupervisor.start_child(ConnectionSupervisor, {Postgrex, args})
     else
-      {:lookup, [{conn, _}]} -> {:ok, conn}
+      {:lookup, [{conn, _, _}]} -> {:ok, conn}
       connection_error -> connection_error
     end
   end
@@ -43,21 +57,17 @@ defmodule Kompost.Kompo.Postgres.Instance do
   Checks in the `Kompost.Kompo.Postgres.ConnectionRegistry` for an existing
   connection defined by the given `id`.
   """
-  @spec lookup(id()) :: [{pid, any}]
-  def lookup(id), do: Registry.lookup(ConnectionRegistry, id)
-
-  @doc """
-  Creates an instance id tuple from a resource.
-
-  ### Example
-
-      iex> resource = %{"metadata" => %{"namespace" => "default", "name" => "foo-bar"}}
-      ...> Kompost.Kompo.Postgres.Instance.get_id(resource)
-      {"default", "foo-bar"}
-  """
-  @spec get_id(resource :: map()) :: id()
-  def get_id(%{"metadata" => metadata}), do: {metadata["namespace"], metadata["name"]}
-  def get_id(reference), do: {reference["namespace"], reference["name"]}
+  @spec lookup(id()) :: [
+          {pid, conn_args :: keyword(),
+           allowed_namespaces :: NamespaceAccess.allowed_namespaces()}
+        ]
+  def lookup(id) do
+    Registry.lookup(ConnectionRegistry, id)
+    |> Enum.map(fn {pid, args} ->
+      {allowed_namespaces, conn_args} = Keyword.pop!(args, :allowed_namespaces)
+      {pid, conn_args, allowed_namespaces}
+    end)
+  end
 
   @doc """
   Checks if the user connected to the given `conn` has all the privileges
@@ -89,7 +99,7 @@ defmodule Kompost.Kompo.Postgres.Instance do
   @spec disconnect(id()) :: :ok
   def disconnect(id) do
     case lookup(id) do
-      [{conn, _}] ->
+      [{conn, _, _}] ->
         # This will also unregister the process at the ConnectionRegistry:
         DynamicSupervisor.terminate_child(ConnectionSupervisor, conn)
         :ok