+++ test/support/kompo/postgres/resource_helper.ex
@@ -4,6 +4,22 @@ defmodule Kompost.Test.Kompo.Postgres.ResourceHelper do
   import Kompost.Test.GlobalResourceHelper
   import YamlElixir.Sigil
 
+  @spec cluster_instance(name :: binary(), opts :: Keyword.t()) :: map()
+  def cluster_instance(name, opts \\ []) do
+    ~y"""
+    apiVersion: kompost.chuge.li/v1alpha1
+    kind: PostgresClusterInstance
+    metadata:
+      name: #{name}
+    spec:
+      hostname: #{System.get_env("POSTGRES_HOST", "127.0.0.1")}
+      port: #{System.fetch_env!("POSTGRES_EXPOSED_PORT")}
+      username: #{System.fetch_env!("POSTGRES_USER")}
+      plainPassword: #{System.fetch_env!("POSTGRES_PASSWORD")}
+    """
+    |> apply_opts(opts)
+  end
+
   @spec instance(name :: binary(), namespace :: binary(), opts :: Keyword.t()) :: map()
   def instance(name, namespace, opts \\ []) do
     ~y"""
@@ -52,24 +68,33 @@ defmodule Kompost.Test.Kompo.Postgres.ResourceHelper do
   @spec database(
           name :: binary(),
           namespace :: binary(),
-          instance :: map(),
+          instance :: {:cluster | :namespaced, map()},
           params :: map(),
           opts :: Keyword.t()
         ) ::
           map()
   def database(name, namespace, instance, params \\ %{}, opts \\ []) do
-    ~y"""
-    apiVersion: kompost.chuge.li/v1alpha1
-    kind: PostgresDatabase
-    metadata:
-      name: #{name}
-      namespace: #{namespace}
-    spec:
-      instanceRef:
-        name: #{instance["metadata"]["name"]}
-        namespace: #{instance["metadata"]["namespace"]}
-    """
-    |> put_in(~w(spec params), params)
-    |> apply_opts(opts)
+    database =
+      ~y"""
+      apiVersion: kompost.chuge.li/v1alpha1
+      kind: PostgresDatabase
+      metadata:
+        name: #{name}
+        namespace: #{namespace}
+      """
+      |> Map.put("spec", %{"params" => params})
+      |> apply_opts(opts)
+
+    case instance do
+      {:cluster, cluster_instance} ->
+        put_in(database, ~w(spec clusterInstanceRef), %{
+          "name" => cluster_instance["metadata"]["name"]
+        })
+
+      {:namespaced, ns_instance} ->
+        put_in(database, ~w(spec instanceRef), %{
+          "name" => ns_instance["metadata"]["name"]
+        })
+    end
   end
 end