+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Changed
+
+- **Breaking:** `PostgresDatabase` - `.spec.instanceRef.namespace` in favor of `PostgresClusterInstance` [#25](https://github.com/mruoss/kompost/pull/25)
+
+### Added
+
+- `PostgresClusterInstance` - Cluster scoped instance accessible by databases in any or a defined set of namespaces. [#25](https://github.com/mruoss/kompost/pull/25)
+
 ## [0.2.3] - 2023-06-30
 
 ### Added